🔥 Remove V2 dual implementation pattern
- Renamed all V2 components back to original names
- Removed V2 suffixes from component exports
- Removed migration indicators
- Fixed imports in page.tsx and test-zustand
- Fixed RecastLogo import path
