+++ temp_page_section.txt
@@ -1,0 +1,6 @@
+        <CodeEditorPanel />
+
+        <CanvasArea />
+
+        <div className="flex">
+          <ControlsPanel />
